Move UpdateExpressionGenerator into its own file
